Create index on nodes and add SSSP method
SSSP (Single Source Shortest Path) was added as a simple method to get the shortest paths given a node name (should be changed to the index instead of name).
The results are still unused. We need to perform this on all nodes, and consider ways to parallelize.
Using the all-pairs-shortest-path built in algorithm in Neo4j takes too much memory as one operation. Even though it has some optimizations, we must divide the computations to multimple single-source calls.